New widget for the Taskbar of the Sketcher WB.
The objectives of this widget are:
1. Allow editing of sketches when they have (partially) overlapping lines of any type
   (construction/external/normal).
2. Allow the construction of sketches using the keyboard (less mouse intensive)

The widget works as follows:
1. A QListWidget presents a list of elements, element is here a geometric element
formed Types edges, starting points, end points and midpoints (e.g. circles, arcs)
2. On hovering/Entering one item in the list, it is preselected for easy identification.
3. The selection can be switched from one type to another using the SHIFT key
(press and release) in a round-robin fashion (or you can select it using the droplist)
4. Selection of different types is possible (e.g. two points and a line).
5. The contextual menu accessible by right click allow to apply a constraint to the
selection.
6. The constraints in the contextual menu have shortcuts, so you can actually use the
shortcut directly from the QListWidget.

Additional features introduced to other parts during development...
- New color introduced for Types that are selected and preselected at the same time
- Edges is a SoLineSet. You can not apply many So modifiers (width,offset) to parts of a SoLineSet.
An edge part of a SoLineSet that is drawn after another one, covers it unless it is closer to the
camera. The system of manual offsets already in place for points has been extended to properly
handle (i.e. show) overlapping lines part of a SoLineSet.