Fixing that a constraint is not added if a constraint having the same first,second,third,firstpos,secondpos,thirdpos and type is existing.
This fixes: http://forum.freecadweb.org/viewtopic.php?f=3&t=6890

i.e. that if three points, where two of them already have a coincident
constraint are selected and a coincident constraint requested, a second
identical constraint is not added (this happens when selecting the three
points with a selection box, if done by clicking, only one of the points of the
already coincident ones is returned).

Side effects:
- It silently prevents from adding the constraint
- When requesting the same constraint with a different value (e.g. Distance constraint)
it will silently ignore the user, which is not nice, it should tell the user about the conflicting
constraints.

Possible solution:
- Constrain the silently ignoring to only coincident constraints.