Sketcher feature enhancement: Symmetry constraint perpendicular to axis of symmetry Solves Mantis ticket: http://www.freecadweb.org/tracker/view.php?id=1623 Unrelated minor fix: remove unused variable
Sketcher visualization enhancement: Do not group sketcher symmetry icons.

Do not add a symmetry constraint icon to a group of icons, so that it lies on the line or point providing the symmetry (symmetry axis/point),
as it is more clear to the user which line this is.

Sketcher Bug fix: Arrows for symmetry constraint are back

The arrows were there in the code, but at a certain point, code was written to set the this->imgHeight value programatically, based on
whether it had text or not. As symmetry constraint does not have text the variable failed to initialize. Being zero, it did not pass
the ASSERT in the render function, so it was not rendered.

Sketcher Art enhancement: New Symmetric constraint icon by Jim

There is more space in between the > < symbols now, so it is more clear as it overlaps less.

Thanks Jim!

Sketcher visualization enhancement: Constraint icons visible over lines