Major fix: Phi definition and Ellipse creation method
Initially the ellipse creation method did not attempt to control whether the user generated a CCW or CW ellipse.

Therefore the ellipse's reference system Z axis sometimes was the opposite of the Z axis of the sketch, which ultimately
led to inconsistent definition of phi, as it was measured with respect to -X (the corresponding X of a right-handed coordinate system with Z in the -Z direction).

Now all the ellipses created by the sketcher are CCW ellipses so that phi is consistent with what the user expects (from where the camera is).

This should fix most problems and inconsistencies with major and minor radii (but yes sometimes still complains that Major<Minor even if not true, this is a separate bug).