Coincident constraint representation: Fix in presence of ellipses and arcsofellipse
The issue:
1. Viewprovidersketch relies on the solver to color the points corresponding to the coincident constraint via getPointId
2. The ellipse and arcofellipse introduce a point (focus1) to the parameters that does not have a visual representation by default
3. getPointId started providing wrong coloring of coincident constraints for every element entered after an ellipse

I think 1 is not ok. It should be decoupled. However, there seems not to be a better way as getVertexIndexGeoPos does not provide the index the color array is expecting ?!

Therefore the solution given here is to make another function getVisiblePointId, that takes into account how many invisible points are before the element and deducts them.